sombot2/requirements.txt
flask
python-dotenv
transformers
vaderSentiment
lime
numpy
google-generativeai
torch
